refactor(pilaf): changing file name to match the Kibbeh name convention for tsx components
