+++ pilaf/src/navigators/bottomNavigator.tsx
@@ -4,9 +4,9 @@ import { DashboardPage } from "../pages/DashboardPage";
 import { ProfilePage } from "../pages/ProfilePage";
 import { SchedulePage } from "../pages/SchedulePage";
 import { FollowingPage } from "../pages/FollowingPage";
-import { colors } from "../constants/GlobalStyles";
+import { colors } from "../constants/dogeStyle";
 import Icon from "react-native-vector-icons/Ionicons";
-import { CreateRoomButton } from "../components/bottomBar/createRoomButton";
+import { CreateRoomButton } from "../components/bottomBar/CreateRoomButton";
 const Tab = createBottomTabNavigator();
 
 const EmptyComponent: React.FC = () => {
