+++ pilaf/src/pages/MainPage.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { Header } from "../components/header";
-import { BottomNavigator } from "../navigators/bottomNavigator";
+import { Header } from "../components/Header";
+import { BottomNavigator } from "../navigators/BottomNavigator";
 
 export const MainPage: React.FC = () => {
   return (
