+++ pilaf/src/pages/ProfilePage.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { StyleSheet, Text } from "react-native";
 import { TouchableOpacity } from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
-import { colors, fontFamily } from "../constants/GlobalStyles";
+import { colors, fontFamily } from "../constants/dogeStyle";
 import { useTokenStore } from "../module/auth/useTokenStore";
 
 export const ProfilePage: React.FC = () => {
