+++ pilaf/src/pages/SchedulePage.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { StyleSheet, Text } from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
-import { colors, fontFamily } from "../constants/GlobalStyles";
+import { colors, fontFamily } from "../constants/dogeStyle";
 
 export const SchedulePage: React.FC = () => {
   return (
